Move Cranelift meeting to Wednesdays.
This lets us avoid the time limits of our now 45-minute Monday slot and
get back to a full hour, which I think will allow us to avoid cutting
off important conversations, or at least better than before.

The calendar event (ping me for an invite!) is still weekly for now,
keeping the overflow-every-other-week that we had after our Monday
timeslot was compressed to 45 minutes; but if folks want to go back to
biweekly we can do that too (happy to discuss in next meeting).

This also has the advantage of avoiding holiday cancelations more often.

I polled the regular attendees on Zulip and this seemed to generally
work but please (as always) let me know if there's an issue.