Add a weekly component model implementation meeting
Some colleagues and I at Fastly have already been meeting at this time
to discuss details about the implementation of the component model in
addition to using it as a coordination point amongst us to ensure we are
all kept up-to-date with what's going on. At the last meeting we thought
it would be good to invite anyone else who's interested in this to also
drop in if they'd like.

Note that this meeting is not about the design of the component model
itself, that's best left for the WASI subgroup meetings. This is only
intended to be for the implementation of the component model as-proposed
and the tooling around it that the Bytecode Alliance is hosting.