Make Cranelift meetings weekly now (and add agendas).
In our meeting today (2022-09-21) we discussed whether it made sense to
formalize the almost-weekly cadence we had been taking (which evolved
from an overflow-into-next-week scheme) into fully weekly. We agreed to
do so, and so this PR adds the remaining agendas for weekly meetings for
the rest of 2022 (except for Dec 28, when I believe most of us will be
on holidays).