Start using agendas for component model meeting
I've enough times wished now for a written down agenda to write on
between weeks that I'm going to go ahead and start this up. I've
deviated from the other templates in this repository by skipping the
notes and attendees for now as I don't want to add too much process yet,
but I figure this can at least serve as the start of having an agenda
anyone can add to.